gke/inference-api/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
xgboost==2.0.3
numpy==1.26.4
google-cloud-storage==2.17.0
pydantic==2.7.1
